+++ cmsafvis/R/warming_stripes_plot.R
@@ -17,7 +17,7 @@
 warming_stripes_plot <- function(variable = NULL,
                                  infile = NULL, 
                                  selected_number = 1,
-                                 stripe_color = 1,
+                                 color_pal = 1,
                                  analyze_method = TRUE,
                                  temp_dir = tempdir(),
                                  out_dir = getwd(),
@@ -182,7 +182,7 @@ warming_stripes_plot <- function(variable = NULL,
   country_name <- get_country_name(country_code)
   # title <- paste0(variable, " Stripes Plot: ", country_name, " (", climate_year_start, " - ", substring(end_date,1,4), ")")
   title <- paste0(variable, " change in ", country_name, " since XXXX (Reference: ", 
-                  climate_year_start, " - ", substring(end_date,1,4), ")")
+                  climate_year_start, " - ", climate_year_end, ")")
   
   plot_warming_stripes(
     variable = variable,
@@ -199,7 +199,7 @@ warming_stripes_plot <- function(variable = NULL,
     pointsTF = pointsTF,
     lineTF = lineTF, 
     title = title,
-    stripe_color = stripe_color,
+    color_pal = color_pal,
     circ_plot = circ_plot,
     verbose = TRUE)
 }