+++ cmsafops/NEWS.md
@@ -1,5 +1,9 @@
 # Changelog for cmsafops package
 
+## 1.2.4
+- Adaptation of time sorting in box_mergetime 
+- Minor bug fix for timcumsum, which might occur in rare cases
+
 ## 1.2.3
 - Bug fix in mon.anomaly: wrong definition of time unit
 - Bug fix in selpoint.multi: time is now double, not integer